service, Controller 수정

+++ src/main/kotlin/com/waffletoy/team1server/post/service/PostService.kt
@@ -4,10 +4,9 @@ import com.waffletoy.team1server.account.controller.User
 import com.waffletoy.team1server.post.PostServiceException
 import com.waffletoy.team1server.post.controller.Post
 import com.waffletoy.team1server.post.controller.PostBrief
-import com.waffletoy.team1server.post.persistence.BookmarkEntity
-import com.waffletoy.team1server.post.persistence.BookmarkRepository
-import com.waffletoy.team1server.post.persistence.PostRepository
-import com.waffletoy.team1server.post.persistence.PostSpecification
+import com.waffletoy.team1server.post.persistence.*
+import org.springframework.data.domain.Page
+import org.springframework.data.domain.PageRequest
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
@@ -31,10 +30,21 @@ class PostService(
         roles: List<String>?,
         investment: Int?,
         investor: List<String>?,
-    ): List<PostBrief> {
-        val specification = PostSpecification.withFilters(roles, investment, investor)
-        val posts = postRepository.findAll(specification)
-        return posts.map { PostBrief.fromPost(Post.fromEntity(it)) }
+        status: Int?,
+        page: Int = 0
+    ): Page<PostEntity> {
+        val specification = PostSpecification.withFilters(
+            roles, investment, investor, status?:2
+        )
+
+        val pageable = PageRequest.of(page, pageSize)
+
+        val postIds = postRepository.findAll(specification).map { it.id }
+
+        // PostEntity를 페이징 처리하여 가져오기
+        return postRepository.findAllByIdIn(postIds, pageable)
+
+//        return posts.map { PostBrief.fromPost(Post.fromEntity(it)) }
     }
 
     @Transactional
@@ -42,6 +52,12 @@ class PostService(
         user: User,
         postId: String,
     ) {
+        if (bookmarkRepository.existsByUserIdAndPostId(user.id, postId)) {
+            throw PostServiceException(
+                "이미 북마크에 추가된 포스트입니다.",
+                HttpStatus.CONFLICT
+            )
+        }
         val post = postRepository.findByIdOrNull(postId) ?: throw PostServiceException(
             "존재하지 않는 채용정보입니다.",
             HttpStatus.NOT_FOUND,
@@ -60,10 +76,12 @@ class PostService(
         user: User,
         postId: String,
     ) {
-        val post = postRepository.findByIdOrNull(postId) ?: throw PostServiceException(
-            "존재하지 않는 채용정보입니다.",
-            HttpStatus.NOT_FOUND,
-        )
+        if (!bookmarkRepository.existsByUserIdAndPostId(user.id, postId)){
+            throw PostServiceException(
+                "존재하지 않는 북마크입니다",
+                HttpStatus.NOT_FOUND,
+            )
+        }
         bookmarkRepository.deleteByUserIdAndPostId(user.id, postId)
         return
     }
@@ -72,21 +90,14 @@ class PostService(
     fun getBookmarks(
         user: User,
         page: Int,
-    ) : List<PostBrief> {
-        val bookmarks = bookmarkRepository.findAllByUserId(user.id)
-            .mapNotNull { bookmark ->
-                postRepository.findByIdOrNull(bookmark.postId) // postId를 사용해 Post 엔티티를 찾음
-            }
+    ) : Page<PostEntity> {
+        val pageable = PageRequest.of(page, pageSize)
 
-        // 페이지네이션 처리
-        val paginatedPosts = bookmarks
-            .drop(page * pageSize) // 현재 페이지까지 데이터를 건너뛰기
-            .take(pageSize)       // 페이지 크기만큼 가져오기
+        // 북마크 ID 가져오기
+        val bookmarkIds = bookmarkRepository.findAllByUserId(user.id).map { it.postId }
 
-        // PostBrief로 변환
-        return paginatedPosts.map { post ->
-            PostBrief.fromPost(Post.fromEntity(post))
-        }
+        // PostEntity를 페이징 처리하여 가져오기
+        return postRepository.findAllByIdIn(bookmarkIds, pageable)
     }
 
     private val pageSize = 12
